server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
git+https://github.com/franco-lamas/PyOBD
